# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  types.AssignTaskPayload:
    properties:
      assigneeId:
        type: integer
    type: object
  types.CreateGoalPayload:
    properties:
      description:
        maxLength: 2000
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - title
    type: object
  types.CreateTaskPayload:
    properties:
      assigneeId:
        type: integer
      description:
        maxLength: 2000
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - title
    type: object
  types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  types.Goal:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      ownerId:
        type: integer
      ownerName:
        type: string
      title:
        type: string
    type: object
  types.GoalWithTasks:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      ownerId:
        type: integer
      ownerName:
        type: string
      tasks:
        items:
          $ref: '#/definitions/types.Task'
        type: array
      title:
        type: string
    type: object
...
  types.Task:
    properties:
      assigneeId:
        type: integer
      assigneeName:
        type: string
      createdAt:
        type: string
      createdBy:
        type: integer
      createdByName:
        type: string
      description:
        type: string
      goalId:
        type: integer
      goalTitle:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
  types.UpdateTaskPayload:
    properties:
      assigneeId:
        type: integer
      description:
        maxLength: 2000
        type: string
      goalId:
        minimum: 1
        type: integer
      status:
        enum:
        - todo
        - in_progress
        - done
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - goalId
    - status
    - title
    type: object
...
  types.UserTasksBoard:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/types.Task'
        type: array
    type: object
info:
  contact: {}
  description: REST API for task tracking with goals and assignments
  title: Task Tracker API
  version: "1.0"
paths:
  /goals:
    get:
      description: Get all goals for the authenticated user with nested tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.GoalWithTasks'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get goals
      tags:
      - goals
    post:
      consumes:
      - application/json
      description: Create a goal owned by the authenticated user
      parameters:
      - description: Goal payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.CreateGoalPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.Goal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create goal
      tags:
      - goals
  /goals/{goalID}:
    delete:
      description: Delete a goal owned by the authenticated user
      parameters:
      - description: Goal ID
        in: path
        name: goalID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete goal
      tags:
      - goals
    put:
      consumes:
      - application/json
      description: Update a goal owned by the authenticated user
      parameters:
      - description: Goal ID
        in: path
        name: goalID
        required: true
        type: integer
      - description: Goal payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.CreateGoalPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Goal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update goal
      tags:
      - goals
  /goals/{goalID}/tasks:
    get:
      description: Get a single goal with its tasks
      parameters:
      - description: Goal ID
        in: path
        name: goalID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GoalWithTasks'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tasks by goal
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a task under a goal. Only the goal owner can create tasks.
      parameters:
      - description: Goal ID
        in: path
        name: goalID
        required: true
        type: integer
      - description: Task payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.CreateTaskPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - tasks
...
  /tasks/{taskID}:
    delete:
      description: Delete a task from a goal owned by the authenticated user
      parameters:
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update a task under a goal owned by the authenticated user
      parameters:
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: integer
      - description: Task update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.UpdateTaskPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - tasks
  /tasks/{taskID}/assign:
    put:
      consumes:
      - application/json
      description: Assign or unassign a task. Only the goal owner can assign tasks.
      parameters:
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: integer
      - description: Assignment payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.AssignTaskPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign task
      tags:
      - tasks
  /tasks/assigned:
    get:
      description: Get tasks assigned to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Task'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get assigned tasks
      tags:
      - tasks
...
  /users/tasks:
    get:
      description: Get all users and their current assigned tasks (todo and in_progress)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.UserTasksBoard'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get users with current tasks
      tags:
      - users
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"